feat(runtime): implement snapshot capture and safe revert
Implement snapshot capture, hashing, and reverse replay with shared filesystem locks, freshness records, containment checks, and cancellation-safe mutation ownership. Qualify the Rust host against TypeScript across all transports and supported platforms.

Preserve cross-device move bytes and permissions, reject occupied destinations before staging, and retain recovery data when publication or restoration cannot be verified. Announce the embedded contract version and keep checkpoint capabilities truthful.

Validated with the full Bun and Rust workspace gates, protocol feature checks and tests, 47 focused filesystem I/O tests, and 24 real-binary Hub qualification tests. Addressed the review findings with deterministic regressions.

+++ crates/mangostudio-runtime/src/filesystem/patch_apply.rs
@@ -183,7 +183,7 @@ fn plan_operation(
                 read_patch_target(service, &policy, &mutation.chat_id, resolved_path, cancel)?
                     .bytes;
             if mutation.capture_snapshot {
-                super::service::snapshot_limit(resolved_path, source.len())?;
+                super::service::snapshot_limit(resolved_path, source.len() as u64)?;
             }
             Ok(PlannedOperation::Delete {
                 input_path: input_path.clone(),
@@ -208,7 +208,7 @@ fn plan_operation(
             let observed =
                 read_patch_target(service, &policy, &mutation.chat_id, resolved_path, cancel)?;
             if mutation.capture_snapshot {
-                super::service::snapshot_limit(resolved_path, observed.bytes.len())?;
+                super::service::snapshot_limit(resolved_path, observed.bytes.len() as u64)?;
             }
             let source = std::str::from_utf8(&observed.bytes).map_err(|_| {
                 io::path_error(format!(
